requirements.txt
python-dotenv==1.2.1
streamlit==1.53.1
requests==2.32.5
db-dtypes==1.5.0
pytz==2025.2
plotly==6.5.2
pyarrow==23.0.0
pandas==2.3.3
pandas-gbq==0.33.0
google-cloud-bigquery==3.40.0
google-genai==1.62.0
pytest==8.3.5
pytest-cov==6.1.0
